reafactor(trash): split to two components
